refactor: base all entities off of a base class
